+++ tests/plugin_test/src/132_compact.php
@@ -1,0 +1,7 @@
+<?php
+function returnsPackedArgs($myVar, $myOtherVar, $myUnusedVar) {
+    $nameToReturn = 'myOtherVar';
+    return compact('myVar', [$nameToReturn]);
+}
+// returns ['myVar' => 'a', 'myOtherVar' => 'b']
+var_export(returnsPackedArgs('a', 'b', 'c'));