+++ NEWS.md
@@ -21,6 +21,7 @@ New features(Analysis):
 + Fix failure to infer type when there is an assignment in a condition (#2964)
   (e.g. `return ($obj = maybeObj()) instanceof stdClass ? $obj : new stdClass();`)
 + Warn about no-ops in for loops (e.g. `for ($x; $x < 10, $x < 20; $x + 1) {}`) (#2926)
++ Treat `compact('var1', ['var2'])` as a usage of $var1 and $var2 in `--unused-variable-detection` (#1812)
 
 Bug fixes:
 + Fix crash in StringUtil seen in php 7.4-dev due to notice in `hexdec()` (affects polyfill/fallback parser).
@@ -30,6 +31,8 @@ Plugins:
   In the `plugin_config` config array, `inline_html_whitelist_regex` and `inline_html_blacklist_regex` can be used to limit the subset of analyzed files to check for inline HTML.
 + For `UnusedSuppressionPlugin`: `'plugin_config' => ['unused_suppression_whitelisted_only' => true]` will make this plugin report unused suppressions only for issues in `whitelist_issue_types`. (#2961)
 + For `UseReturnValuePlugin`: warn about unused results of function calls in loops (#2926)
++ Provide the `$node` causing the call as a 5th parameter to closures returned by `AnalyzeFunctionCallCapability->getAnalyzeFunctionCallClosuresStatic`
+  (this can be used to get the variable/expression for an instance method call, etc.)
 
 Maintenance:
 + Made `--polyfill-parse-all-element-doc-comments` a no-op, it was only needed for compatibility with running Phan with php 7.0.